Added `sf::Packet` Utility class and added testing impl for PubSub
The current implementation for the Publish Subscribe System will not
work completely due to the lack of run-time casting in C++, therefore we
can not send a dynamic `std::any` to a function, determine it's type,
and cast to that type. This will need to be re-thought.

This commit however adds a utility class that allows `sf::Event`'s' to be
sent through `sf::Socket`'s'.