Event pub sub (#1)
* Implemented base parts and beggining of PubSub

Kind of a late commit, oh well.

This commit introduces a basic format for code and files.
This commit also starts to implement an Event Based Publish Subscribe.
The Publish Subscribe classes will ultimately use UDP sockets to send
the messages, probably using SFML UDP sockets.

See [this](https://github.com/proatgram/Notes/blob/master/Projects/Sole/Event%20Based%20Publisher%20and%20Subscriber.md) for design information.

* Added `sf::Packet` Utility class and added testing impl for PubSub

The current implementation for the Publish Subscribe System will not
work completely due to the lack of run-time casting in C++, therefore we
can not send a dynamic `std::any` to a function, determine it's type,
and cast to that type. This will need to be re-thought.

This commit however adds a utility class that allows `sf::Event`'s' to be
sent through `sf::Socket`'s'.

* Started implementing Event class

This `Sole::Core::Event` class will handle all comparisons against their
respective types.

* Implemented Event class

This class is used as an event itself, or can be used as a container for
a bunch of different event types.

* Removed Event PubSub system

This system was a neat addition, but it wasn't the thing that was needed
for a "game" this early on in development. Thus it has been removed, to
allow focus to be on things that would add to the game more. Such as,
movement?